+++ src/theme.ts
@@ -9,6 +9,9 @@ export const theme = {
     mainColor: 'hsl(256, 64%, 62%)',
     mainColorAccent: 'hsl(256, 69%, 69%)',
     mainColorShadow: 'hsla(256, 64%, 62%, 0.3)',
+    secondaryColor: 'hsl(150, 64%, 62%)',
+    secondaryColorAccent: 'hsl(150, 69%, 69%)',
+    secondaryColorShadow: 'hsla(150, 64%, 62%, 0.3)',
     bgColor: 'hsl(255, 100%, 100%)',
     maxResponsiveWidth: '800px',
 }